Move the modules to the correct location (#3462)
